refactor: extract admin docs endpoints into dedicated router
